move api.views to api.viewsets for better consistency
