+++ cloudigrade/api/tests/viewsets/test_machineimageviewset.py
@@ -7,7 +7,7 @@ from rest_framework.test import APIRequestFactory, force_authenticate
 from api.clouds.aws.models import AwsMachineImage
 from api.models import MachineImage
 from api.tests import helper as api_helper
-from api.views import MachineImageViewSet
+from api.viewsets import MachineImageViewSet
 from util.tests import helper as util_helper
